requirements.txt
rich==13.7.1
rich-argparse==1.4.0
datasets==2.19.0
emoji==2.11.1
parallel-pandas==0.6.3
tabulate==0.9.0
